requirements/base.txt
Django==4.2.4
djangorestframework==3.14.0
dj-rest-auth==2.2.4
drf-spectacular==0.26.4
requests==2.28.0
psycopg2-binary==2.9.7
urllib3==1.26.9
redis==5.0.0
django-redis==5.3.0
celery==5.3.4